Add test:arel to test separately from adapters
Previously, all of the Arel tests would be run with every database
adapter. This is not necessarily a problem, but these tests end up
running redundantly for each adapter/database combination even though
they do not interact with adapters at all.

This commit follows up a [commit][1] in Rails that added a new test:arel
task for Active Record. This additional step creates a place for Arel to
be tested a single time, so that a followup PR to Rails can filter out
Arel tests when testing adapters.

Since the task is only present on the main branch, it cannot run for all
Rails versions and must be a special case.

[1] rails/rails@f362f0796f2dd750fc03892ed45a57d145426bff