Skip checking out rails/rails in test steps
While it appears to be cached for some steps, the checkout takes 20-30
seconds when it doesn't cache. Since all of the tests run inside docker
images the checkout is unnecessary so it can be disabled to save the
time.

When enabling this setting, Buildkite gets more strict over how the
Docker images are built. Because of this, we have to provide the other
compose service names in the "build-alias" field.