Remove guide testing for Ruby 2.7 and Ruby 3.0
Now that main has been updated to only support Ruby 3.1+, the guides
tests on 7-0-stable and 7-1-stable are failing for Ruby 2.7 and 3.0
(since half the bug report templates are for the main branch). This
commit removes the steps to test guides on these Ruby versions so that
there are no longer failing steps on stable branches.

Alternatively, we could backport the bug report template refactors that
removed the template pointing at main in favor of a comment inside the
gem template.