Don't pull compose images
That pull take 30-40s per job, which is quite insane when these
image rarely if ever change.

Of course if they significantly change it would cause problem,
but then we should probably find another way to deal with this
like running a specific tag that is committed in this repo
and that we manually update once in a while.