Remove bundle install from docs-preview pipeline
The reason for `bundle install` was because the initial PR we used the `ruby:latest` docker image, which doesn't have any dependencies pre-installed.

As such, we needed `BUNDLE_WITHOUT` to prevent installing gems which depend on native extensions also missing from that image.

Since we switched to using a pre-built mainline image, all of the necessary dependencies should be there and we can remove this step to save time.