Add Rack 3.0 jobs
Since the release of Rack 3.1, we are now only testing against Rack 2
and Rack 3.1. This is a problem because their are behavioral changes
made between 3.0 and 3.1 (such as `rack.input` now being optional).

To ensure that we stay compatible with all of the versions we claim to
be compatible with, this commit adds actionpack and railties jobs for
Rack 3.0.